+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.2.18 (unreleased)
 -------------------
 
+- WEB-4366 : Add RAM caching to @search endpoint. Try to reduce latency
+  [boulch]
+
 - Migrate to Plone 6.1.3
   [boulch]